# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.createLeaseRequest:
    properties:
      brand_id:
        type: integer
      building_id:
        type: integer
      customer_id:
        type: integer
      department_id:
        type: integer
      end_date:
        type: string
      lease_no:
        type: string
      management_type_id:
        type: integer
      start_date:
        type: string
      store_id:
        type: integer
      tenant_name:
        type: string
      terms:
        items:
          $ref: '#/definitions/handlers.createLeaseTermRequest'
        type: array
      trade_id:
        type: integer
      units:
        items:
          $ref: '#/definitions/handlers.createLeaseUnitRequest'
        type: array
    required:
    - department_id
    - end_date
    - lease_no
    - start_date
    - store_id
    - tenant_name
    - terms
    - units
    type: object
  handlers.createLeaseTermRequest:
    properties:
      amount:
        type: number
      billing_cycle:
        $ref: '#/definitions/lease.BillingCycle'
      currency_type_id:
        type: integer
      effective_from:
        type: string
      effective_to:
        type: string
      term_type:
        $ref: '#/definitions/lease.TermType'
    required:
    - amount
    - billing_cycle
    - currency_type_id
    - effective_from
    - effective_to
    - term_type
    type: object
  handlers.createLeaseUnitRequest:
    properties:
      rent_area:
        type: number
      unit_id:
        type: integer
    required:
    - rent_area
    - unit_id
    type: object
...
  handlers.submitLeaseRequest:
    properties:
      comment:
        type: string
      idempotency_key:
        type: string
    required:
    - idempotency_key
    type: object
  handlers.swaggerEnvelope:
    type: object
  handlers.swaggerMessageResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  handlers.terminateLeaseRequest:
    properties:
      terminated_at:
        type: string
    required:
    - terminated_at
    type: object
...
  lease.BillingCycle:
    enum:
    - monthly
    type: string
    x-enum-varnames:
    - BillingCycleMonthly
  lease.Contract:
    properties:
      amended_from_id:
        type: integer
      approved_at:
        type: string
      billing_effective_at:
        type: string
      brand_id:
        type: integer
      building_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      customer_id:
        type: integer
      department_id:
        type: integer
      effective_version:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      lease_no:
        type: string
      management_type_id:
        type: integer
      start_date:
        type: string
      status:
        $ref: '#/definitions/lease.Status'
      store_id:
        type: integer
      submitted_at:
        type: string
      tenant_name:
        type: string
      terminated_at:
        type: string
      terms:
        items:
          $ref: '#/definitions/lease.Term'
        type: array
      trade_id:
        type: integer
      units:
        items:
          $ref: '#/definitions/lease.Unit'
        type: array
      updated_at:
        type: string
      updated_by:
        type: integer
      workflow_instance_id:
        type: integer
    type: object
  lease.Status:
    enum:
    - draft
    - pending_approval
    - active
    - rejected
    - terminated
    type: string
    x-enum-varnames:
    - StatusDraft
    - StatusPendingApproval
    - StatusActive
    - StatusRejected
    - StatusTerminated
  lease.Summary:
    properties:
      billing_effective_at:
        type: string
      brand_id:
        type: integer
      building_id:
        type: integer
      customer_id:
        type: integer
      department_id:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      lease_no:
        type: string
      management_type_id:
        type: integer
      start_date:
        type: string
      status:
        $ref: '#/definitions/lease.Status'
      store_id:
        type: integer
      tenant_name:
        type: string
      trade_id:
        type: integer
      updated_at:
        type: string
      workflow_instance_id:
        type: integer
    type: object
  lease.Term:
    properties:
      amount:
        type: number
      billing_cycle:
        $ref: '#/definitions/lease.BillingCycle'
      created_at:
        type: string
      currency_type_id:
        type: integer
      effective_from:
        type: string
      effective_to:
        type: string
      id:
        type: integer
      lease_contract_id:
        type: integer
      term_type:
        $ref: '#/definitions/lease.TermType'
      updated_at:
        type: string
    type: object
  lease.TermType:
    enum:
    - rent
    - deposit
    type: string
    x-enum-varnames:
    - TermTypeRent
    - TermTypeDeposit
  lease.Unit:
    properties:
      created_at:
        type: string
      id:
        type: integer
      lease_contract_id:
        type: integer
      rent_area:
        type: number
      unit_id:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:5180
info:
  contact: {}
  description: Backend API for the MI property management system
  title: MI Management Information System API
  version: "1.0"
paths:
...
  /leases:
    get:
      consumes:
      - application/json
      description: Returns paginated lease summaries filtered by lease number, status,
        and store.
      parameters:
      - description: Lease number
        in: query
        name: lease_no
        type: string
      - description: Lease status
        in: query
        name: status
        type: string
      - description: Store ID
        in: query
        name: store_id
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/lease.Summary'
                  type: array
                page:
                  type: integer
                page_size:
                  type: integer
                total:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: List leases
      tags:
      - Lease
    post:
      consumes:
      - application/json
      description: Creates a draft lease contract with units and billing terms.
      parameters:
      - description: Lease create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.createLeaseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                lease:
                  $ref: '#/definitions/lease.Contract'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Create lease draft
      tags:
      - Lease
  /leases/{id}:
    get:
      consumes:
      - application/json
      description: Returns a lease contract with units and terms by ID.
      parameters:
      - description: Lease ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                lease:
                  $ref: '#/definitions/lease.Contract'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Get lease
      tags:
      - Lease
  /leases/{id}/amend:
    post:
      consumes:
      - application/json
      description: Creates an amended lease draft from an existing lease contract.
      parameters:
      - description: Lease ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lease amendment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.createLeaseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                lease:
                  $ref: '#/definitions/lease.Contract'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Create lease amendment
      tags:
      - Lease
  /leases/{id}/submit:
    post:
      consumes:
      - application/json
      description: Submits a draft lease contract into the workflow approval process.
      parameters:
      - description: Lease ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lease submit request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.submitLeaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                lease:
                  $ref: '#/definitions/lease.Contract'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Submit lease for approval
      tags:
      - Lease
  /leases/{id}/terminate:
    post:
      consumes:
      - application/json
      description: Terminates an active lease contract on the supplied termination
        date.
      parameters:
      - description: Lease ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lease termination request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.terminateLeaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.swaggerEnvelope'
            - properties:
                lease:
                  $ref: '#/definitions/lease.Contract'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.swaggerMessageResponse'
      security:
      - BearerAuth: []
      summary: Terminate lease
      tags:
      - Lease
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"